+++ chatwoot/java/src/main/java/oseg/chatwoot_examples/ContactInboxCreationExample.java
@@ -8,6 +8,7 @@ import com.chatwoot.client.JSON;
 import com.chatwoot.client.model.*;
 
 import java.io.File;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.OffsetDateTime;
 import java.util.ArrayList;
@@ -24,13 +25,13 @@ public class ContactInboxCreationExample
         // ((ApiKeyAuth) config.getAuthentication("platformAppApiKey")).setApiKey("PLATFORM_APP_API_KEY");
 
         var contactInboxCreationRequest = new ContactInboxCreationRequest();
-        contactInboxCreationRequest.inboxId(0);
+        contactInboxCreationRequest.inboxId(BigDecimal.valueOf(0));
 
         try
         {
             var response = new ContactApi(config).contactInboxCreation(
                 0, // accountId
-                0, // id
+                BigDecimal.valueOf(0), // id
                 contactInboxCreationRequest // data
             );
