+++ chatwoot/java/src/main/java/oseg/chatwoot_examples/ContactUpdateExample.java
@@ -8,6 +8,7 @@ import com.chatwoot.client.JSON;
 import com.chatwoot.client.model.*;
 
 import java.io.File;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.OffsetDateTime;
 import java.util.ArrayList;
@@ -29,7 +30,7 @@ public class ContactUpdateExample
         {
             var response = new ContactsApi(config).contactUpdate(
                 0, // accountId
-                0, // id
+                BigDecimal.valueOf(0), // id
                 contactUpdate // data
             );
