+++ chatwoot/java/src/main/java/oseg/chatwoot_examples/UpdateInboxExample.java
@@ -8,6 +8,7 @@ import com.chatwoot.client.JSON;
 import com.chatwoot.client.model.*;
 
 import java.io.File;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.OffsetDateTime;
 import java.util.ArrayList;
@@ -30,7 +31,7 @@ public class UpdateInboxExample
         {
             var response = new InboxesApi(config).updateInbox(
                 0, // accountId
-                0, // id
+                BigDecimal.valueOf(0), // id
                 updateInboxRequest // data
             );
